Special Character Handling
Added the ability to handle whitespace, special characters and
capital letters. Special characters, including whitespace, are
passed through untouched without advancing the rotors. Capital
letters are converted to lowercase, passed through, and then
recapitalized.

To Be Added:
Ability to handle multi-line input as well as encryption of and
decryption to plain text files on local machine.

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 Changes committed:
	modified:   EnigPy.py