Multi-line Input
Added ability to handle multi-line input terminated by a blank line
without any blank lines including through a clever use of list()
and iter() (Considering changing to sys.stdin.read() so blank lines
may be included). Using pathlib the script can now read in and
delete a file for encryption/decryption and write its output to a
file along with the previously existing options.

To Be Added:
Change from a procedural to an object oriented style.

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 Changes committed:
	modified:   .gitignore
	modified:   EnigPy.py
	new file:   fileIO.py