User Input Handling
Changed the handling of user input so blank lines can be entered.
Added tabs and newline characters to special character handling.
The encryption will respect all formatting but if will always
correctly decrypt message if formatting is not used when
decryptiing cypher text.

To Be Added:
GUI
Allow for saving/loading of multiple machines based on some kind
of user id.

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 Changes committed:
	modified:   EnigmaMachine.py