HTML Template
Created the HTML template (using bootstrap 4) for the web interface.
There is not yet any functionality, this is only a template. Design
may be tweaked slightly in the future but the look will be more or
less constant from this point forward.

To Be Added:
Allow user to enter plain/cypher text and handle passing that to
an instance of EnigmaMachine and then updating page with encrypted/
decrypted text.

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 Changes committed:
	modified:   templates/home.html