Function Web Interface
A web interface with basic function:
	pass plain text message
	receive encrypted message
	pass encrypted message
	receive original plain text
has been implemented. The user is not currently able to set the
rotor settings so they may decrypt messages aside from the most
recently encrypted message. The rotor settings are set as a global
variable and updated upon each encryption. A flag is passed when
rendering an html template so messages aren't flashed in two places
at once. This is a poor implementation and should be fixed in future
commits. Some sections of EnigmaMachine.py were commented where
functionality was not needed or complicated basic implementation.
It was also noticed that the program up to this point cannot handle
numbers. This functionality will be implemented into the master branch
and in a future commit on this branch as well.
A virtual environment (called EnigPy) was created in the repo.

To Be Added:
User ability to set machine settings
Ability to display encryption/decryption path
Ability to create nodes so users can communicate
with machines set up the same way

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 On branch flask
 Changes committed:
	modified:   EnigmaMachine.py
	modified:   Rotors.py
	new file:   forms.py
	modified:   routes.py
	modified:   templates/home.html